Register atlas textures with texture manager

+++ src/main/java/xyz/flirora/caxton/font/CaxtonInternal.java
@@ -19,7 +19,7 @@ public class CaxtonInternal {
 
     public static native int fontAtlasNumPages(long addr);
 
-    public static native int fontAtlasPage(long addr, int pageNum);
+    public static native long fontAtlasPage(long addr, int pageNum);
 
     public static native ShapingResult[] shape(long fontAddr, char[] s, int[] bidiRuns);
 }
